requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
google-auth>=2.28,<3.0
google-auth-oauthlib>=1.2,<2.0
google-api-python-client>=2.120,<3.0
openai>=1.0,<2.0
whitenoise>=6.6,<7.0
gunicorn>=21.2,<23.0
